backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
# psycopg2-binary==2.9.10   will install for prod (using sqlite for dev)

# SQLite support (for dev/testing)
aiosqlite==0.20.0

# Pydantic (FastAPI uses this for schemas)
pydantic==2.12.4
pydantic-settings==2.6.1

# Google Sheets API
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-cov==6.0.0

# Dev utilities
python-dotenv==1.0.1